backend/backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
mssql-django==1.6
PyJWT==2.10.1
pyodbc==5.3.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
